+++ src/bin/actions/build-number.action.ts
@@ -1,5 +1,5 @@
 import { Action } from "./action";
-import { lastCommitHasTag, git, repoHasLocalCommits } from "../../lib/git";
+import { git } from "../../lib/git";
 import { getPackageVersion, getVersionFile, setPackageVersion } from "../../lib/version-file";
 import { ReleaseType } from "../../lib/version";
 import { logger } from "../../lib/util";
@@ -12,32 +12,16 @@ export class BuildNumberAction implements Action {
     name = "incrementBuildNumber";
 
     /** The description of the action */
-    description = "Increments the build number by one and appends it to the last commit.";
+    description = "Increments the build number by one.";
 
     /**
      * Run the action
      */
     async run(): Promise<void> {
-        if (await lastCommitHasTag()) {
-            logger.error(
-                "You cannot increase the build number of an already tagged commit. This would change its hash and the tag won't match the right commit anymore.",
-            );
-            return;
-        }
-
-        // Only execute if there are local commits
-        if (!(await repoHasLocalCommits())) {
-            logger.warn(
-                "No local commits detected. The --incrementBuildNumber action should only be executed if your repository has local commits to which the changes can be appended to. It is recommended to automatically execute the action in a post commit hook.",
-            );
-            return;
-        }
-
         const version = await getPackageVersion();
         version.increase(ReleaseType.build);
         await setPackageVersion(version);
         await git.add(getVersionFile());
-        await git.commit([], { "--amend": null, "--no-edit": null });
         logger.info(`Version was incremented to ${version.toString()}. The update was appended to the last commit.`);
     }
 }