+++ tests/bin/actions/init.action.spec.ts
@@ -93,8 +93,8 @@ describe("init action integration test", () => {
                 "\n" +
                 "export const configuration: Partial<WersionConfigModel> = {\n" +
                 "  versionFile: {\n" +
-                '      path: `./package.json`,\n' +
-                '      matcher: `"version": ?"${semverMatcher}"`,\n' +
+                "      path: `<enter file>`,\n" +
+                '      matcher: `"version": ?"${semverMatcher}"`\n' +
                 "  },\n" +
                 "  commitTypes: {\n" +
                 "      major: [],\n" +
@@ -103,7 +103,30 @@ describe("init action integration test", () => {
                 "  },\n" +
                 '  breakingChangeTrigger: "breaking change",\n' +
                 '  changelogFilePath: "./CHANGELOG.md",\n' +
-                '  projectName: "wersion-test",\n' +
+                '  projectName: "wersion-test"\n' +
+                "};",
+        );
+    });
+
+    it("should use the right configuration for custom preset", async () => {
+        const action = new InitAction();
+        const template = action.compileWersionRCTsTemplate({ preset: "Custom", projectName: "wersion-test" });
+        expect(template).toEqual(
+            'import { WersionConfigModel, semverMatcher } from "@weichwarenprojekt/wersion";\n' +
+                "\n" +
+                "export const configuration: Partial<WersionConfigModel> = {\n" +
+                "  versionFile: {\n" +
+                "      path: `<enter file>`,\n" +
+                '      matcher: `"version": ?"${semverMatcher}"`\n' +
+                "  },\n" +
+                "  commitTypes: {\n" +
+                "      major: [],\n" +
+                '      minor: ["feat"],\n' +
+                '      patch: ["fix"]\n' +
+                "  },\n" +
+                '  breakingChangeTrigger: "breaking change",\n' +
+                '  changelogFilePath: "./CHANGELOG.md",\n' +
+                '  projectName: "wersion-test"\n' +
                 "};",
         );
     });
@@ -116,8 +139,8 @@ describe("init action integration test", () => {
                 "\n" +
                 "export const configuration: Partial<WersionConfigModel> = {\n" +
                 "  versionFile: {\n" +
-                '      path: `./package.json`,\n' +
-                '      matcher: `"version": ?"${semverMatcher}"`,\n' +
+                "      path: `./package.json`,\n" +
+                '      matcher: `"version": ?"${semverMatcher}"`\n' +
                 "  },\n" +
                 "  commitTypes: {\n" +
                 "      major: [],\n" +
@@ -126,7 +149,7 @@ describe("init action integration test", () => {
                 "  },\n" +
                 '  breakingChangeTrigger: "breaking change",\n' +
                 '  changelogFilePath: "./CHANGELOG.md",\n' +
-                '  projectName: "wersion-test",\n' +
+                '  projectName: "wersion-test"\n' +
                 "};",
         );
     });
@@ -139,8 +162,8 @@ describe("init action integration test", () => {
                 "\n" +
                 "export const configuration: Partial<WersionConfigModel> = {\n" +
                 "  versionFile: {\n" +
-                '      path: `./pubspec.yaml`,\n' +
-                '      matcher: `version: ?${semverMatcher}`,\n' +
+                "      path: `./pubspec.yaml`,\n" +
+                "      matcher: `version: ?${semverMatcher}`\n" +
                 "  },\n" +
                 "  commitTypes: {\n" +
                 "      major: [],\n" +
@@ -149,7 +172,7 @@ describe("init action integration test", () => {
                 "  },\n" +
                 '  breakingChangeTrigger: "breaking change",\n' +
                 '  changelogFilePath: "./CHANGELOG.md",\n' +
-                '  projectName: "wersion-test",\n' +
+                '  projectName: "wersion-test"\n' +
                 "};",
         );
     });