+++ vite.config.ts
@@ -6,5 +6,9 @@ export default defineConfig({
             inline: ["@weichwarenprojekt/ts-importer"],
         },
         exclude: [...defaultExclude],
+        coverage: {
+            // you can include other reporters, but 'json-summary' is required, json is recommended
+            reporter: ["text", "json-summary", "json"],
+        },
     },
 });